+++ primitives/runtime-interface/test-wasm-deprecated/Cargo.toml
@@ -19,7 +19,7 @@ sp-io = { version = "2.0.0", default-features = false, path = "../../io" }
 sp-core = { version = "2.0.0", default-features = false, path = "../../core" }
 
 [build-dependencies]
-wasm-builder-runner = { version = "2.0.0", package = "substrate-wasm-builder-runner", path = "../../../utils/wasm-builder-runner" }
+substrate-wasm-builder = { version = "3.0.0", path = "../../../utils/wasm-builder" }
 
 [features]
 default = [ "std" ]
