+++ utils/wasm-builder-runner/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "substrate-wasm-builder-runner"
-version = "2.0.0"
+version = "3.0.0"
 authors = ["Parity Technologies <admin@parity.io>"]
 description = "Runner for substrate-wasm-builder"
 edition = "2018"
